requirements.txt
aliyun-python-sdk-core-v3==2.9.4
aliyun-python-sdk-dysmsapi==1.0.0
aliyunsdkcore==1.0.3
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
Babel==2.5.3
backports-abc==0.5
beautifulsoup4==4.6.0
billiard==3.5.0.3
cached-property==1.5.1
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.3
django-autofixture==0.12.1
django-celery==3.2.2
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-filter==1.1.0
django-guardian==1.4.9
django-ipware==2.0.1
django-jsonfield==1.0.1
django-ranged-response==0.2.0
django-redis==4.8.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.2.0
django-simple-captcha==0.5.6
djangorestframework==3.7.7
djangorestframework-camel-case==0.2.0
et-xmlfile==1.0.1
ffmpy==0.2.2
flower==0.9.2
geetest==3.2.1
geopy==1.11.0
html5lib==1.0.1
idna==2.5
iso8601==0.1.12
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
jpush==3.2.9
kombu==4.1.0
m3u8==0.3.5
Markdown==2.6.11
MarkupSafe==1.0
mongoengine==0.15.0
mysqlclient==1.3.14
olefile==0.45.1
openapi-codec==1.3.2
Pillow==5.3.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.2
pymongo==3.6.0
PyMySQL==0.9.2
python-alipay-sdk==1.6.0
python-dateutil==2.7.5
pytz==2017.2
PyYAML==3.13
qiniu==7.2.2
raven==6.10.0
redis==2.10.5
requests==2.18.2
simplejson==3.11.1
six==1.10.0
tokenlib==2.0.0
tornado==4.5.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
websocket-client==0.54.0
xlrd==1.1.0
xlwt==1.3.0
